# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PushEventOut": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "title": "PushEventOut",
        "type": "object"
      },
      "PushEventsIn": {
        "properties": {
          "enabled": {
            "items": {
              "type": "string"
            },
            "title": "Enabled",
            "type": "array"
          },
          "endpoint": {
            "default": "",
            "title": "Endpoint",
            "type": "string"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "PushEventsIn",
        "type": "object"
      },
      "PushEventsOut": {
        "properties": {
          "available": {
            "items": {
              "$ref": "#/components/schemas/PushEventOut"
            },
            "title": "Available",
            "type": "array"
          },
          "device": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "enabled": {
            "items": {
              "type": "string"
            },
            "title": "Enabled",
            "type": "array"
          }
        },
        "required": [
          "available",
          "enabled"
        ],
        "title": "PushEventsOut",
        "type": "object"
      },
      "PushRulesIn": {
        "properties": {
          "quiet_end": {
            "default": "",
            "title": "Quiet End",
            "type": "string"
          },
          "quiet_start": {
            "default": "",
            "title": "Quiet Start",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "default": {},
            "title": "Tags",
            "type": "object"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          }
        },
        "title": "PushRulesIn",
        "type": "object"
      },
      "PushRulesOut": {
        "properties": {
          "quiet_end": {
            "default": "",
            "title": "Quiet End",
            "type": "string"
          },
          "quiet_now": {
            "default": false,
            "title": "Quiet Now",
            "type": "boolean"
          },
          "quiet_start": {
            "default": "",
            "title": "Quiet Start",
            "type": "string"
          },
          "tags": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "default": {},
            "title": "Tags",
            "type": "object"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          }
        },
        "title": "PushRulesOut",
        "type": "object"
      },
      "PushSubscribeIn": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "subscription": {
            "additionalProperties": true,
            "title": "Subscription",
            "type": "object"
          }
        },
        "required": [
          "subscription"
        ],
        "title": "PushSubscribeIn",
        "type": "object"
      },
      "PushTestIn": {
        "properties": {
          "endpoint": {
            "default": "",
            "title": "Endpoint",
            "type": "string"
          }
        },
        "title": "PushTestIn",
        "type": "object"
      },
      "PushTestOut": {
        "properties": {
          "sent": {
            "title": "Sent",
            "type": "integer"
          }
        },
        "required": [
          "sent"
        ],
        "title": "PushTestOut",
        "type": "object"
      },
...
      "RestoreIn": {
        "properties": {
          "credentials": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Credentials",
            "type": "array"
          },
          "kv": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Kv",
            "type": "object"
          },
          "push_subscriptions": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Push Subscriptions",
            "type": "array"
          },
          "services": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "default": {},
            "title": "Services",
            "type": "object"
          },
          "stats_samples": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Stats Samples",
            "type": "array"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "title": "RestoreIn",
        "type": "object"
      },
      "RestoreOut": {
        "properties": {
          "credentials": {
            "default": 0,
            "title": "Credentials",
            "type": "integer"
          },
          "kv": {
            "default": 0,
            "title": "Kv",
            "type": "integer"
          },
          "push_subscriptions": {
            "default": 0,
            "title": "Push Subscriptions",
            "type": "integer"
          },
          "services": {
            "default": 0,
            "title": "Services",
            "type": "integer"
          },
          "stats": {
            "default": 0,
            "title": "Stats",
            "type": "integer"
          }
        },
        "title": "RestoreOut",
        "type": "object"
      },
...
      "ServiceSettingsIn": {
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "password": {
            "default": "",
            "title": "Password",
            "type": "string"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          },
          "username": {
            "default": "",
            "title": "Username",
            "type": "string"
          }
        },
        "title": "ServiceSettingsIn",
        "type": "object"
      },
...
      "SettingsImportIn": {
        "properties": {
          "services": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Services",
            "type": "object"
          }
        },
        "required": [
          "services"
        ],
        "title": "SettingsImportIn",
        "type": "object"
      },
...
      "StatsSampleOut": {
        "properties": {
          "disk_free_bytes": {
            "default": 0,
            "title": "Disk Free Bytes",
            "type": "integer"
          },
          "episode_files": {
            "default": 0,
            "title": "Episode Files",
            "type": "integer"
          },
          "indexer_grabs": {
            "default": 0,
            "title": "Indexer Grabs",
            "type": "integer"
          },
          "indexer_queries": {
            "default": 0,
            "title": "Indexer Queries",
            "type": "integer"
          },
          "library_bytes": {
            "default": 0,
            "title": "Library Bytes",
            "type": "integer"
          },
          "movies": {
            "default": 0,
            "title": "Movies",
            "type": "integer"
          },
          "series": {
            "default": 0,
            "title": "Series",
            "type": "integer"
          },
          "torrents_qbit": {
            "default": 0,
            "title": "Torrents Qbit",
            "type": "integer"
          },
          "torrents_tm": {
            "default": 0,
            "title": "Torrents Tm",
            "type": "integer"
          },
          "ts": {
            "title": "Ts",
            "type": "integer"
          }
        },
        "required": [
          "ts"
        ],
        "title": "StatsSampleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WebhookAppOut": {
        "properties": {
          "app": {
            "title": "App",
            "type": "string"
          },
          "configured": {
            "default": false,
            "title": "Configured",
            "type": "boolean"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "installed": {
            "default": false,
            "title": "Installed",
            "type": "boolean"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "app"
        ],
        "title": "WebhookAppOut",
        "type": "object"
      },
      "WebhookInstallIn": {
        "properties": {
          "base_url": {
            "title": "Base Url",
            "type": "string"
          }
        },
        "required": [
          "base_url"
        ],
        "title": "WebhookInstallIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "arrdeck",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/push/events": {
      "get": {
        "operationId": "push_events_api_v1_push_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "endpoint",
            "required": false,
            "schema": {
              "default": "",
              "title": "Endpoint",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushEventsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Events",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "description": "With a subscribed endpoint the choice is that device's alone; otherwise\nit becomes the default that unconfigured devices follow.",
        "operationId": "save_push_events_api_v1_push_events_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushEventsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushEventsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Push Events",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/push/rules": {
      "get": {
        "operationId": "push_rules_api_v1_push_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushRulesOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Push Rules",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "save_push_rules_api_v1_push_rules_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRulesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushRulesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Push Rules",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/push/subscribe": {
      "post": {
        "operationId": "push_subscribe_api_v1_push_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscribeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Subscribe",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/push/test": {
      "post": {
        "operationId": "push_test_api_v1_push_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushTestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Test",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/push/unsubscribe": {
      "post": {
        "operationId": "push_unsubscribe_api_v1_push_unsubscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscribeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Unsubscribe",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/push/webhook/install": {
      "post": {
        "operationId": "webhook_install_api_v1_push_webhook_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookInstallIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebhookAppOut"
                  },
                  "title": "Response Webhook Install Api V1 Push Webhook Install Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Webhook Install",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/restore": {
      "post": {
        "operationId": "restore_api_v1_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/settings/import": {
      "post": {
        "operationId": "import_settings_api_v1_settings_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/settings/services/{name}": {
      "put": {
        "operationId": "save_settings_api_v1_settings_services__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Settings Api V1 Settings Services  Name  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/services/{name}/test": {
      "post": {
        "operationId": "test_service_api_v1_settings_services__name__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Test Service Api V1 Settings Services  Name  Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Service",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/stats/history": {
      "get": {
        "operationId": "stats_history_api_v1_stats_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StatsSampleOut"
                  },
                  "title": "Response Stats History Api V1 Stats History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stats History",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}